% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{find_children}
\alias{find_children}
\title{Recursive helper function for graph traversal}
\usage{
find_children(nodes, edges)
}
\arguments{
\item{nodes}{starting node(s)}

\item{edges}{edgelist (data.frame: from and to)}
}
\value{
vector with all_children from nodes
}
\description{
Recursive helper function for graph traversal
}
