% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{intersect_from_list}
\alias{intersect_from_list}
\title{Helper function for list intersection}
\usage{
intersect_from_list(input)
}
\arguments{
\item{input}{list of gene sets (or any sets)}
}
\value{
binary matrix of elements x sets
}
\description{
Adapted from : UpSetR::fromList but extended with rownames
}
