% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{median_rank}
\alias{median_rank}
\title{Median rank across a matrix of gene expression}
\usage{
median_rank(expression_matrix, subset = NULL, convert = TRUE)
}
\arguments{
\item{expression_matrix}{expression_matrix a matrix or dataframe with rows as voxels and columns as genes}

\item{subset}{numeric vector of ids to subset matrix to. defaults to NULL (no subset)}

\item{convert}{do 1 - (expression_matrix_rank_median / nrow(expression_matrix)) ?}
}
\value{
vector with median ranks per voxel and numeric ids as names
}
\description{
Median rank across a matrix of gene expression
}
