% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{summariseRegions_genesets}
\alias{summariseRegions_genesets}
\title{Annotate sets of cluster Markers with regions}
\usage{
summariseRegions_genesets(gene_set, findRegion_result)
}
\arguments{
\item{gene_set}{gene set(s)}

\item{findRegion_result}{result list from findRegions_genesets}
}
\value{
vector with all_children from nodes
}
\description{
TODO: summarise to dataframe function:
need function that returns likely and other clusters as data.frame
}
