% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_functions.R
\name{aba_ids}
\alias{aba_ids}
\title{retrieve aba ids for a vector of genes}
\usage{
aba_ids(genes, aba_gene_to_id = NULL)
}
\arguments{
\item{genes}{a vector of gene symbols}

\item{aba_gene_to_id}{a dataframe ('id','gene_symbol'), a filepath to a valid dataframe or NULL. if NULL: will use cocoframer to query}
}
\value{
vector with ids and genes as names
}
\description{
retreives all ids associated with provided ids. Genes that are not included in aba will be silently dropped!
}
