% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_functions.R
\name{coco_expression}
\alias{coco_expression}
\title{Retrieve aba expression data per voxel for given ids}
\usage{
coco_expression(
  ids,
  return = "matrix",
  aba_ish_list = NULL,
  aba_ish_matrix = NULL
)
}
\arguments{
\item{ids}{a vector of aba ids}

\item{return}{should the function return list or matrix (see below)}

\item{aba_ish_list}{a list of 3d arrays or a string to a cached version (.rds)}

\item{aba_ish_matrix}{a voxel x gene matrix or a string to a cached version (table)}
}
\value{
a list or matrix depending on return
}
\description{
Retrieves expression data per voxel.
}
