% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{findRegions_genesets}
\alias{findRegions_genesets}
\title{Annotate sets of cluster Markers with regions}
\usage{
findRegions_genesets(
  gene_set,
  aba_gene_to_id = NULL,
  aba_ish_matrix = NULL,
  ...
)
}
\arguments{
\item{gene_set}{gene set(s)}

\item{aba_gene_to_id}{..}

\item{aba_ish_matrix}{..}

\item{...}{parameters pass down}
}
\value{
vector with all_children from nodes
}
\description{
TODO:need to allow confidence measure based on detected genes
}
