% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{findRegions_genesets}
\alias{findRegions_genesets}
\title{Annotate sets of genes (cluster Markers) with mouse brain regions}
\usage{
findRegions_genesets(
  gene_set,
  gene_set_weights = NULL,
  min_ids = 0,
  aba_gene_to_id = NULL,
  aba_ish_matrix = NULL,
  max_ids_to_include = Inf,
  ...
)
}
\arguments{
\item{gene_set}{gene set(s). A list of character vectors}

\item{min_ids}{for quality control (pct_of_genes): How many ids should be included in the calculation. else pct_of_genes will be 0}

\item{aba_gene_to_id}{a dataframe ('id','gene_symbol'), a filepath to a valid dataframe that can be read with data.table::fread or NULL. if NULL: will use cocoframer::get_gene_aba_ish_ids to query}

\item{aba_ish_matrix}{a voxel x gene matrix or a string to a cached version (table) If NULL, cocoframer::get_aba_ish_data will be used to retrieve 3 arrays for all provided ids. If a matrix is provided it will always return a matrix.}

\item{max_ids_to_include}{maximum of ids that will bes used fr enrichment. defaults to inf}

\item{...}{parameters passed down to \code{\link{region_annotation}}}
}
\value{
a list with results
}
\description{
TODO: need to explain all results in details
}
