% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{region_annotation}
\alias{region_annotation}
\title{Annotate regions via voxels using ABA ISH expression data and an enrichment function.}
\usage{
region_annotation(
  expression_matrix,
  target_structure_id = "1097",
  target_level = "8",
  enrich_function = "median_rank",
  topn = Inf,
  aba_ccf_grid_annotation = NULL,
  mba_ontology_flatten = NULL
)
}
\arguments{
\item{expression_matrix}{expression_matrix a matrix or dataframe with rows as voxels and columns as genes}

\item{target_structure_id}{defaults to 1097 for hypothalamus. set to 997 for root!}

\item{target_level}{number st_level in mba data (e.g. level 8 to provide a summary on this level)}

\item{enrich_function}{a function (defaults to 'median_rank') for enrichment}

\item{topn}{topn voxels to annotate region}

\item{aba_ccf_grid_annotation}{a 3d array from cocoframer::get_ccf_grid_annotation() or a filepath to a rds containing a cached version. NULL will query the API}

\item{mba_ontology_flatten}{a dataframe from  cocoframer::get_mba_ontology() + cocoframer::flatten_mba_ontology() or a filepath to a tsv containing a cached version. NULL will query the API}
}
\value{
vector with median ranks per voxel and numeric ids as names
}
\description{
Function uses column names of aba ish objects as provided when queried using cocoframer.
}
