% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{summariseRegions_genesets}
\alias{summariseRegions_genesets}
\title{Annotate sets of cluster Markers with regions}
\usage{
summariseRegions_genesets(findRegion_result, min_score = 0.8)
}
\arguments{
\item{findRegion_result}{result list from findRegions_genesets}

\item{min_score}{min_score to filter for in findRegion_result enrichment (standard function return 0-1, and the default is set to 0.8)}
}
\value{
vector with all_children from nodes
}
\description{
Annotate sets of cluster Markers with regions
}
